+++ playground/ta/parser_test/Makefile
@@ -1,0 +1,46 @@
+NAME		:= a.out
+
+CC			:= cc
+CFLAGS		:= -Wall -Wextra -Werror
+
+ifdef SANI
+	CFLAGS	+= -g -fsanitized=address
+endif
+
+SRCS		= main.c \
+			  tokenizer.c \
+			  parser.c
+
+OBJ_DIR		:= obj
+OBJ			:= $(SRCS:%.c=%.o)
+OBJS		:= $(addprefix $(OBJ_DIR)/, $(OBJ))
+
+LIBFT_DIR	:= ../../../libft
+LIBFT		:= $(LIBFT_DIR)/libft.a
+
+.PHONY	:= all
+all	:	$(NAME)
+
+$(NAME)	: $(OBJS)
+	@make -C $(LIBFT_DIR)
+	$(CC) $(CFLAGS) -o $@ $(LIBFT) $^
+
+$(OBJ_DIR)/%.o	: %.c
+	@mkdir -p $$(dirname $@)
+	$(CC) $(CFLAGS) -o $@ $< -c
+
+.PHONY	:= clean
+clean	:
+	rm -rf $(OBJ_DIR)
+
+.PHONY	:= fclean
+fclean	: clean
+	$(RM) $(NAME)
+
+.PHONY	:= re
+re	: fclean
+	@make all
+
+.PHONY	:= run
+run	: all
+	./$(NAME)
